Loading indicator: refactor
