--- public/app/features/logs/components/log-context/types.ts
@@ -1,1 +1,0 @@
-export type Place = 'above' | 'below';
